requirements.txt
praw
pymongo
python-dotenv
pyspark